+++ flowcraft/generator/components/metagenomics.py
@@ -380,3 +380,41 @@ class Mafft(Process):
         self.status_channels = [
             "mafft"
         ]
+
+class Raxml(Process):
+    """mafft to align sequences
+
+            This process is set with:
+
+                - ``input_type``: align
+                - ``output_type``: .tree
+                - ``ptype``: tree
+
+            """
+
+    def __init__(self, **kwargs):
+
+        super().__init__(**kwargs)
+
+        self.input_type = "align"
+        self.output_type = ".tree"
+
+        self.params = {
+            "substitutionModel": {
+                "default": "'GTRGAMMA'",
+                "description": "Substitution model. Option: GTRCAT, GTRCATI, ASC_GTRCAT, GTRGAMMA, ASC_GTRGAMMA etc "
+            }
+        }
+
+        self.directives = {
+            "raxml": {
+                "container": "flowcraft/raxml",
+                "version": "8.2.11-1",
+                "cpus": 4,
+                "memory": "{ 4.GB * task.attempt }"
+            }
+        }
+
+        self.status_channels = [
+            "raxml"
+        ]